+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/batch/BatchStrategy.java
@@ -20,12 +20,17 @@ package com.axelor.apps.supplychain.service.batch;
 import com.axelor.apps.base.db.repo.BatchRepository;
 import com.axelor.apps.base.service.administration.AbstractBatch;
 import com.axelor.apps.sale.db.SaleOrder;
+import com.axelor.apps.stock.db.StockMove;
 import com.axelor.apps.supplychain.service.SaleOrderInvoiceService;
 import com.axelor.inject.Beans;
 
 public abstract class BatchStrategy extends AbstractBatch {
 
   protected SaleOrderInvoiceService saleOrderInvoiceService;
+
+  protected BatchStrategy() {
+    super();
+  }
 
   protected BatchStrategy(SaleOrderInvoiceService saleOrderInvoiceService) {
     super();
@@ -35,6 +40,13 @@ public abstract class BatchStrategy extends AbstractBatch {
   protected void updateSaleOrder(SaleOrder saleOrder) {
 
     saleOrder.addBatchSetItem(Beans.get(BatchRepository.class).find(batch.getId()));
+
+    incrementDone();
+  }
+
+  protected void updateStockMove(StockMove stockMove) {
+
+    stockMove.addBatchSetItem(Beans.get(BatchRepository.class).find(batch.getId()));
 
     incrementDone();
   }